refactor font handling and clear font with buffer
also changes font command format so command byte comes before bufferId, as for some future commands buffer ID will be optional
