+++ video/vdu_buffered.h
@@ -7,6 +7,7 @@
 #include <unordered_map>
 
 #include "agon.h"
+#include "agon_fonts.h"
 #include "buffers.h"
 #include "buffer_stream.h"
 #include "mem_helpers.h"
@@ -295,6 +296,7 @@ void VDUStreamProcessor::bufferClear(uint16_t bufferId) {
 	if (bufferId == 65535) {
 		buffers.clear();
 		resetBitmaps();
+		resetFonts();
 		resetSamples();
 		return;
 	}
@@ -305,6 +307,7 @@ void VDUStreamProcessor::bufferClear(uint16_t bufferId) {
 	}
 	buffers.erase(bufferIter);
 	clearBitmap(bufferId);
+	clearFont(bufferId);
 	clearSample(bufferId);
 	debug_log("bufferClear: cleared buffer %d\n\r", bufferId);
 }
@@ -1188,6 +1191,7 @@ void clearTarget(uint16_t target) {
 		bufferIter->second.clear();
 	}
 	clearBitmap(target);
+	clearFont(target);
 }
 
 void clearTargets(tcb::span<const uint16_t> targets) {
